Revert "Revert "Refactor of styling""
