+++ src/components/ColorModeToggler/styles.module.scss
@@ -1,3 +1,5 @@
+@use '../../css/breakpoints.scss' as breakpoints;
+
 .colorModeButton {
   display: flex;
   align-items: center;
@@ -19,20 +21,5 @@
   white-space: nowrap;
 }
 
-@media (max-width: 1100px) {
-  .themeModeText {
-    display: none;
-  }
-}
 
-@media (max-width: 997px) {
-  .themeModeText {
-    display: block;
-  }
-
-  .colorModeWrapper {
-    display: flex;
-    gap: 0.3rem;
-  }
-}
