+++ src/components/MobileSideBarMenu/Hamburger.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './styles.module.css'
+import styles from './styles.module.scss'
 
 const Hamburger = ({onClick}) =>
 {
