+++ src/components/MobileSideBarMenu/styles.module.scss
@@ -1,5 +1,7 @@
+@use '../../css/breakpoints.scss' as breakpoints;
+
 .docsNavBurger {
-    display: none;
+    display: flex;
 }
 
 .docsMobileMenu {
@@ -24,9 +26,14 @@
 }
 
 .docsMobileMenuHidden {
+    display: none;
     position:absolute;
     opacity: 0;
     transform: translate3d(-100%, 0, 0);
+}
+
+.docsMobileMenuBackdropInactive{
+
 }
 
 [data-theme="dark"] .docsMobileMenuBackdropActive{
@@ -63,6 +70,7 @@
     margin-top: 1em;
     margin-bottom: 2em;
 }
+
 .docsMobileMenu_header{
     margin-top: 4rem;
     margin-bottom: 1rem;
@@ -72,8 +80,8 @@
     justify-content: space-between;
 }
 
-@media (max-width: 996px) {
+@media (min-width: breakpoints.$laptop-breakpoint) {
     .docsNavBurger {
-        display: flex;
+        display: none;
     }
 }
