+++ src/css/breakpoints.css
@@ -1,0 +1,10 @@
+/* Breakpoints for responsive design
+ * These variables can be used in media queries to apply styles at different screen sizes.
+ */
+
+:root {
+  --mobile-breakpoint: 996px;
+  --tablet-breakpoint: 1024px;
+  --laptop-breakpoint: 1330px;
+  --large-desktop-breakpoint: 1440px;
+}
