+++ src/css/custom.scss
@@ -962,7 +962,7 @@ prism-code {
   }
 }
 
-@media screen and (max-width: 1100px) {
+@media screen and (max-width: var(--laptop-breakpoint)) {
   .dropdown {
     padding-right: 2px;
     padding-left: 20px;
