+++ src/css/home.scss
@@ -272,4 +272,3 @@ img.home-svg svg path {
   max-width: 1000px;
   justify-self: center;
 }
-
