+++ src/theme/BlogSidebar/Desktop/styles.module.css
@@ -49,7 +49,7 @@
   margin-bottom: 1.75rem;
 }
 
-@media (max-width: 996px) {
+@media (max-width: var(--mobile-breakpoint)) {
   .sidebar {
     display: none;
   }
