+++ src/theme/DocItem/Footer/styles.module.css
@@ -4,7 +4,7 @@
   font-size: smaller;
 }
 
-@media (min-width: 997px) {
+@media (min-width: var(--mobile-breakpoint)) {
   .lastUpdated {
     text-align: right;
   }
