+++ src/theme/DocRoot/Layout/Main/styles.module.css
@@ -5,7 +5,7 @@
   padding-bottom: 2rem;
 }
 
-@media (min-width: 997px) {
+@media (min-width: var(--mobile-breakpoint)) {
   .docMainContainer {
     flex-grow: 1;
     max-width: calc(100% - var(--doc-sidebar-width));
