+++ src/theme/DocSidebar/Desktop/Content/styles.module.css
@@ -1,4 +1,4 @@
-@media (min-width: 997px) {
+@media (min-width: var(--mobile-breakpoint)) {
   .menu {
     flex-grow: 1;
     padding: 0 1rem 0.5rem;
