+++ src/theme/Footer/Layout/styles.module.scss
@@ -1,6 +1,8 @@
+@use '../../../css/breakpoints.scss' as breakpoints;
+
 .footer {
+  display: none;
   border: 1px solid var(--click-color-stroke);
-  display: flex;
   flex-direction: column-reverse;
   align-items: center;
   gap: 16px;
@@ -25,8 +27,9 @@
   width: 2.5rem;
 }
 
-@media (min-width: 768px) {
+@media (min-width: breakpoints.$tablet-breakpoint) {
   .footer {
+    display: flex;
     /* padding: 32px 40px; */
     padding: 0;
     flex-direction: row;
@@ -40,10 +43,6 @@
     justify-content: space-between;
     margin: 16px;
   }
-}
 
-@media (max-width: 768px) {
-  .footer {
-      display: none;
-  }
+
 }
