+++ src/theme/Navbar/Content/GlobalMenu/styles.module.scss
@@ -1,3 +1,5 @@
+@use '../../../../css/breakpoints.scss' as breakpoints;
+
 .linkStyle {
   display: block;
   color: var(--ifm-navbar-link-color);
@@ -99,7 +101,7 @@
   font-weight: 500;
 }
 
-@media screen and (min-width: 640px) {
+@media screen and (min-width: breakpoints.$mobile-breakpoint) {
   .menuItemContainer {
     min-width: 500px;
   }
@@ -108,7 +110,7 @@
   }
 }
 
-@media screen and (min-width: 1024px) {
+@media screen and (min-width: breakpoints.$tablet-breakpoint) {
   .subMenuItem {
     min-width: 9.5rem;
   }
@@ -186,7 +188,7 @@
  padding-bottom: 10px;
 }
 
-@media only screen and (min-width: 600px) {
+@media only screen and (min-width: breakpoints.$mobile-breakpoint) {
   .navigationMenuContent {
     width: auto;
   }
@@ -218,7 +220,7 @@
   filter: drop-shadow(0px 2px 2px rgba(50, 50, 50, 0.1));
 }
 
-@media only screen and (min-width: 600px) {
+@media only screen and (min-width: breakpoints.$mobile-breakpoint) {
   .navigationMenuViewport {
     width: var(--radix-navigation-menu-viewport-width);
   }
@@ -326,10 +328,4 @@
   to {
     opacity: 0;
   }
-}
-
-@media (min-width: 997px) {
- .navigationMenuRoot {
-   display: flex;
- }
 }
