+++ src/theme/SearchPage/styles.module.css
@@ -55,7 +55,7 @@
   font-style: italic;
 }
 
-@media only screen and (max-width: 996px) {
+@media only screen and (max-width: var(--mobile-breakpoint)) {
   .searchQueryColumn {
     max-width: 60% !important;
   }
